完善了 ustl::vector 整体架构

+++ include/config.h
@@ -8,8 +8,8 @@ namespace ustl
 
 #if __ustl_lib == 20221018UL
     typedef     char                byte;
-    typedef     long long           diff_t;
-    typedef     unsigned long long  size_t;
+    typedef     long                diff_t;
+    typedef     unsigned long       size_t;
     typedef     double              float_t;
     typedef     void *              void_ptr;
